# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OmniGest",
        "description": "Documentação da API Omnigest",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://sistema-generico-api.vercel.app/api/api"
        }
    ],
    "paths": {
...
        "/metodoPagamento": {
            "post": {
                "tags": [
                    "MetodoPagamento"
                ],
                "summary": "Cria um novo método de pagamento",
                "operationId": "1b76c69b3e79454bc442e2198d1bc171",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "desc_metodo_pagamento_tmp"
                                ],
                                "properties": {
                                    "desc_metodo_pagamento_tmp": {
                                        "type": "string",
                                        "example": "Cartão de Crédito"
                                    },
                                    "is_ativo_tmp": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "id_empresa_tmp": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Método de pagamento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MaterialMovimentacao"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação dos dados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro de validação"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/metodoPagamento/{id_metodo_pagamento}": {
            "get": {
                "tags": [
                    "MetodoPagamento"
                ],
                "summary": "Obtém um método de pagamento específico",
                "operationId": "706784495870a0ab0c18aac9f41c148c",
                "parameters": [
                    {
                        "name": "id_metodo_pagamento",
                        "in": "path",
                        "description": "ID do método de pagamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Método de pagamento encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MaterialMovimentacao"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Método de pagamento não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Metodo de Pagamento Não Existe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "MetodoPagamento"
                ],
                "summary": "Atualiza um método de pagamento existente",
                "operationId": "5ec176ab3c111a2ef8ec7df27333f61a",
                "parameters": [
                    {
                        "name": "id_metodo_pagamento",
                        "in": "path",
                        "description": "ID do método de pagamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "desc_metodo_pagamento_tmp"
                                ],
                                "properties": {
                                    "desc_metodo_pagamento_tmp": {
                                        "type": "string",
                                        "example": "Cartão de Débito"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Método de pagamento atualizado com sucesso"
                    },
                    "400": {
                        "description": "Erro na validação dos dados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro de validação"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "MetodoPagamento"
                ],
                "summary": "Deleta um método de pagamento",
                "operationId": "cc94c3f73c0d67be3e9bdd82b10b6e99",
                "parameters": [
                    {
                        "name": "id_metodo_pagamento",
                        "in": "path",
                        "description": "ID do método de pagamento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Método de pagamento deletado com sucesso"
                    },
                    "404": {
                        "description": "Método de pagamento não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Método de pagamento não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MaterialMovimentacao": {
                "properties": {
                    "id_movimentacao_item_mit": {
                        "type": "integer"
                    },
                    "id_movimentacao_mit": {
                        "description": "Chave estrangeira referenciando a movimentação",
                        "type": "integer"
                    },
                    "id_empresa_mit": {
                        "description": "Chave estrangeira referenciando a empresa",
                        "type": "integer"
                    },
                    "id_material_mit": {
                        "description": "Chave estrangeira referenciando o material",
                        "type": "integer"
                    },
                    "qtd_material_mit": {
                        "description": "Quantidade do material movimentado",
                        "type": "integer"
                    },
                    "is_ativo_mit": {
                        "description": "Indica se o item está ativo",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    },
                    "tipo_movimentacao_mit": {
                        "description": "Tipo de movimentação do item",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}